Refactor to using a JobProperty with upgrade test.
